refactor: reorganize into modular project structure
Move flat scripts into organized modules:
- arena/ — core trading logic (models, tools, strategies, dashboard, recorder)
- deploy/ — deployment scripts (chat_node, router_node, tools_and_dashboard, response_viewer)
- exchanges/ — exchange connectors (binance, coinbase)
- tests/ — test suite

Remove start_arena.py and update docs to reflect manual component startup.
