+++ deploy/router_node.py
@@ -1,0 +1,105 @@
+"""Deploy a single named AgentRouterNode for the daytrading arena.
+
+Each router subscribes to the shared ``agent_router.input`` topic with its
+own consumer group, so every agent receives every market tick independently.
+The ``--chat-node-name`` flag targets a specific named ChatNode for LLM
+inference.
+
+Example:
+    uv run python deploy/router_node.py \
+        --name momentum --chat-node-name gpt5-nano --strategy momentum \
+        --bootstrap-servers <broker-url>
+
+    uv run python deploy/router_node.py \
+        --name brainrot-daytrader --chat-node-name deepseek --strategy brainrot \
+        --bootstrap-servers <broker-url>
+"""
+
+import argparse
+import asyncio
+import sys
+
+from calfkit.broker.broker import BrokerClient
+from calfkit.nodes.agent_router_node import AgentRouterNode
+from calfkit.nodes.chat_node import ChatNode
+from calfkit.runners.service import NodesService
+from calfkit.stores.in_memory import InMemoryMessageHistoryStore
+from arena.tools import calculator, execute_trade, get_portfolio
+from arena.strategies import STRATEGIES
+
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Deploy a named AgentRouterNode for the daytrading arena.",
+    )
+    parser.add_argument(
+        "--name",
+        required=True,
+        help="Agent name (used as consumer group + identity)",
+    )
+    parser.add_argument(
+        "--chat-node-name",
+        required=True,
+        help="Name of the deployed ChatNode to target (e.g. gpt5-nano)",
+    )
+    parser.add_argument(
+        "--strategy",
+        required=True,
+        choices=list(STRATEGIES.keys()),
+        help="Trading strategy (selects system prompt)",
+    )
+    parser.add_argument(
+        "--bootstrap-servers",
+        required=True,
+        help="Kafka bootstrap servers address",
+    )
+    return parser.parse_args()
+
+
+async def main() -> None:
+    args = parse_args()
+
+    system_prompt = STRATEGIES.get(args.strategy)
+    if system_prompt is None:
+        print(f"ERROR: Unknown strategy '{args.strategy}'")
+        print(f"Available: {', '.join(STRATEGIES.keys())}")
+        sys.exit(1)
+
+    print("=" * 50)
+    print(f"Router Node Deployment: {args.name}")
+    print("=" * 50)
+
+    print(f"\nConnecting to Kafka broker at {args.bootstrap_servers}...")
+    broker = BrokerClient(bootstrap_servers=args.bootstrap_servers)
+    service = NodesService(broker)
+
+    # ChatNode reference for topic routing (deployed separately via deploy/chat_node.py)
+    chat_node = ChatNode(name=args.chat_node_name)
+
+    tools = [execute_trade, get_portfolio, calculator]
+    router = AgentRouterNode(
+        chat_node=chat_node,
+        tool_nodes=tools,
+        name=args.name,
+        message_history_store=InMemoryMessageHistoryStore(),
+        system_prompt=system_prompt,
+    )
+    service.register_node(router, group_id=args.name)
+
+    tool_names = ", ".join(t.tool_schema.name for t in tools)
+    print(f"  - Agent:    {args.name}")
+    print(f"  - Strategy: {args.strategy}")
+    print(f"  - ChatNode: {args.chat_node_name} (topic: {chat_node.entrypoint_topic})")
+    print(f"  - Input:    {router.subscribed_topic}")
+    print(f"  - Reply:    {router.entrypoint_topic}")
+    print(f"  - Tools:    {tool_names}")
+
+    print("\nRouter node ready. Waiting for requests...")
+    await service.run()
+
+
+if __name__ == "__main__":
+    try:
+        asyncio.run(main())
+    except KeyboardInterrupt:
+        print("\nRouter node stopped.")
